+++ library/irmc_raid.py
@@ -11,17 +11,15 @@ module: irmc_raid
 short_description: handle iRMC RAID
 
 description:
-    - This module has not been verified on iRMC S6. Verification is planned for a future version.
     - Ansible module to configure a PRIMERGY server's RAID via iRMC.
     - Using this module may force the server into several reboots.
-    - Module Version V1.2.
+    - Module Version V1.3.0.
 
 requirements:
     - The module needs to run locally.
-    - The PRIMERGY server needs to be at least a M2 model.
-    - iRMC S4 needs FW >= 9.04, iRMC S5 needs FW >= 1.25.
-    - Python >= 2.6
-    - Python modules 'future', 'requests', 'urllib3'
+    - iRMC S6
+    - Python >= 3.10
+    - Python modules 'requests', 'urllib3'
 
 version_added: "2.4"
 
@@ -49,6 +47,8 @@ options:
         choices:     ['get', 'create', 'delete']
     adapter:
         description: The logical number of the adapter to create/delete RAID arrays on/from.
+                     The logical number is the value at the end of the id (ex. “RAIDAdapter0”)
+                     obtained by command=“get”.
         required:    false
     array:
         description: The logical number of the RAID array to delete. Use -1 for all arrays. Ignored for 'create'.
@@ -83,7 +83,6 @@ EXAMPLES = r'''
       ansible.builtin.debug:
         var: raid.configuration
 
-# Create RAID array
 - name: Create RAID array
   irmc_raid:
     irmc_url: "{{ inventory_hostname }}"
@@ -98,7 +97,6 @@ EXAMPLES = r'''
   tags:
     - create
 
-# Delete RAID array
 - name: Delete RAID array
   irmc_raid:
     irmc_url: "{{ inventory_hostname }}"